Create the Iniital NIF structure
Here we create a wrapper for the rust project Taiga, and create an
Rust-->Erlang NIF so we can call our Taiga code from Elixir

A lot more needs to happen on the Rust side. Sadly this means we will
be needing to fork `n` dependencies, where `n` is the number of crates
the types of Taiga relies on.

Rust is quite fun....