Move Anoma.Node to Anoma.Node.Executor
The layout of Anoma.Node is ambiguious, as what is a the Primary Node?
We have specific kinds of nodes, like intent pool nodes, memory pool
nodes, etc.

To make it explicit what kind of node it is, we've moved it to be an
Executor node, which is the closest node it can be.

What is there is a stub and should change, as an Executor node does a
lot more than what is there