Primary And Communicator Nodes properly fulfill their role
Before the modules were mimicking what an IntentPool would be.

Now we have properly casted them into being a barebones exectuor node
that just runs the logic functions.

In the future this node will expand to creating nullfiiers and
commitment keys as a result of executing the logic

In doing so, we remove the existing tests, as they are no longer
relevant to the logic