Update Anoma.PartialTx.is_valid/1 into a proper calling convention
Before `Anoma.PartialTx.is_valid/1` would crash upon calling, as the
nested lists were not managed properly.

Now the interface was updated to be somewhat sensicle.

Anoma.Eval.Apply Instead of taking a `Anoma.PartialTx.t()`, we instead
take an `Anoma.Eval.input()`. which is a triple of:

1. if it's an input or output
2. a flattened list of inputs
3. a flattened list of outputs

Thus, the new logic properly flattens the list and gives it to the
rpoper call