ci: also run on automerge with next
Try an automerge with next in parallel, and also run checks on the
automerge result.

If the automerge fails, pass the job; there's no way I'm able to find
to skip it or give it a neutral result instead. The desired behavior is
to only display the red X on things worth bringing to an author's
attention; the maintainer will know to run tests manually after
resolving a conflict.