ci: have all checks run even if one fails
Assuming setup reached any of the actual checks, they should all run,
not just until one fails.

I did write a parallel multi-job version of this CI pipeline, but am
not using it because the tests run so quickly that this just runs
slower and spends more CI minutes. It may be revisited if the test
suite ever grows slow.