Setup Phoenix
Setup Phoenix Web functionality for anoma/anoma. This is done without
establishing an umbrella project. Instead, we used mix to create a
separate phoenix `anoma` project in anpther directory and manually
added novel files as well as changed the dependencies in the mix file.

The original project was created with options `--no-ecto` and `--live`

Updates the README file with instructions for server launch.