Change Intent Communicator and Pool Functionality
Changes pool GenServer casts to calls for submitting and removing
intents. Corresponding functions now return info on whether the intent
has been submitted before or whether there is no corresponding intent
to remove and does nothing to the pool in this case. Based on this,
changes the Communicator casts to broadcast intent removal and
addition if pool was changed.