Replace IO.inspect/1 and IO.inspect/2 calls with Logger calls
This replaces our adhoc IO.inspect with a instrumentation flag, to
just using the inbuilt logging functionality.

A lot of the calls we had got moved to Logger.info since most of them
are informative rather than excessive. However some did get moved to
warning and debug

The warning is for failed workers, this happens rarly enough that it
shouldn't flood the user's REPL, and the debug calls are the more
excessive bits of information that isn't helpful