Create An automated communicator that can be `use/2`'d
Here we create a communicator that can be `use/2`'d. Meaning that
instead of having to by hand deal with the details of subscription and
un-subscription, this can mostly be automated.

In the future we can upgrade the type to listen to certain kinds of
topics, meaning that we can do more elaborate strategies as time
progresses with this module automating the annoying bits. Without
having to redo this x times in a row