Change test_helper/nock.ex to be mounted from hoon directly
Previously the file hoon/tests.hoon gave instructions on how to load
it. It assumed inputting it into the dojo directly. Instead we decide
to make it properly loadable like all other files in the Hoon
directly.

A guide on a seperate branch shows how to do this, so it makes
onboarding this easily.

Further since I made it rely on `resource-machine` rather than
`anoma`, we have changed the `stdlib_core()` into `logics_core()` and
have updated the references to be inline with the new changes.