Create an index for ExDoc generation
We are able to do this by abusing symlinks. Sadly this means for
Windows the generation may not be optimal, we should do this in an OS
agnostic way.

Further we update every index link to go back to the index_docs.

We should maybe consider removing `index.livebook` entirely, however
I'm unsure if we should go against convention here