Replace all require: true from typedstruct with enforce: true
It seems that typedstruct and field from Typedstruct does not error
upon getting keys which it does not know about. This has left our code
to have some errors hidden inside of it.

This topic replaces all those wiht the proper enforce.

Further we fixup new compiler errors that stem from this issue as
well.

Namely the mempool tries to make it's map by using an empty primary,
so the enforced fields have to be removed