Create the verificaiton engine
This topic creates the verification engine.

Note that we don't implement `signs_for_lookup`, as currently there is
no signs_for engine. This will eventually be implemented and replace
the default false value constructed.

_This topic overall respects specs except in 1 spot_:

Namely we wish to take a table for the `signsFor` argument, as for
testing I'm unsure of which `signsFor` table we would even consult.