Create tests for generating and connecting to identities
We simply try out the functionality and make sure the engines we
generate work properly. Note it is important to try the
Commitment.commit(com, 555), as otherwise we don't know if they are
setup correctly.

Further we create tests for making sure the proper permissions are
respected, I.E. generating with :commit vs :commit_and_decrypt