Add a common public and secret key to the crypto modules
This will be useful as in the future we may wish to add new public and
secret keys. And thus I want a stable type, that will be a sum type of
all known cryptographic signing and encryption keys.

In the future we may break up this module into more sub signing and
sub encrypted modules, with the module that is now wrapping over it,
changing it's API slightly