Update the commitment engine to have combined and detatched mode.
Further we update the tests to match the new API

One decision we have made is that the data is automatically turned
into a binary if it isn't already. This policy decision should prevent
accidental errors where binary is turned into a binary and now has
failed, or the user forgets and the system crashes as it expects it to
be binary.

Check at the source idempotentcy basically.