Add Home Directory Configuration
Adds Anoma Home Directory configuration to the Anoma configuration
file. We follow XDG format and create directories for data, config,
and state appropriately, either by reading from the environment or by
setting defaults.

The directory names are created as `anoma_name` where `name` is the
name of the environment Anoma was launched in.

Sets mnesia directory to the data directory.