Reformatting the Storage Engine
Reformats the storage engine using a section format similar to that of
the Mempool. The sections are as follows:

- Public RPC API
Contains all Router calls and casts functions showcasing user API

- GenServer Behavior
Contains all the handles for the call and casts of the RPC API

- GenServer Implementation
Contains all functions directly used in the definitions GenServer
Behavior functions.

- Blocking + Snapshots
Contain isolated blocking and snapshotting functionality which is not
used in any other sections.

- Coceptual Helpers
Contain the functions directly used in the definitions in the above
sections

- Helpers
Contain the rest of the functions. Mainly private functions used in
the definitions of Conceptual Helpers as well as `check_if_any_absent`
being an unsused function.

In every section, the law governing function ordering is: if `foo`
uses `bar` in its definition then `foo` is placed higher than
`bar`. This is generally true of the overall stucture of the Engine
modulo `checked_read_at`