Add the elixir equivalent functions for important bitwise operations
These functions correspond to the ones tested in the last commit.

This commit does not register the jet functions, but merely creates
equivalent functions to the non jetted version.

We will in a future commit test these by putting these in the jet
list, effectively using the previous commit's test as a regression
test