Move out Bitwise fiddling helper functions to it's own Module
This is important as we have 3 modules (soon) that will be using
various helper functions around bytes. And having each of those
modules call into each other for bitwise fidling is confusing.

This helps centralize the functions we need to do the implementation,
and provides a clear place for future helper functions.

squash