Add Logging Topic
Adds a logger topic functionality for the logger. Before the way of
using the logger was limited to simply getting the messages our of the
storage and printing them out. Now the user can simply listen to the
logger and flush whenever they would want to get the info on the most
recent actions.