Pad signature binary to 64 bytes when decoding
Noun.atom_integer_to_binary decodes an integer to binary in the minimal
number of bytes required. So we need to specify the expected size of the
binary when decoding.

In this case it is possible that a 64 byte signature may be encoded to
an integer which will fit in 63 or fewer bytes.