Create TestMacro Module
Creates a TestHelper.TestMacro module. Its main functionality is in
replasing ExUnit usage throughout the testing codebase.

Our `__using__` functions similarly to ExUnit.Case but does not import
Assertions module. Instead, we provide our own assertions which
default to the Assertions module macros unless we are in the debug
environment, in which case we pry opon catching any errors.