Add a testing flag to configuration
This makes it so that any test with this flag won't try to take over
the default port and cause TCP listeners to crash (not a big deal,
just noisy).

This is part 1 of the changes. another part will cover loading a dump,
but another topic is needed first before that can be tackled