Normalize noun before updating jam cache
The jam implmentation uses a cache to record positions of jammed nouns.
This is used to identify and write backrefs.

A normalized representation of a noun must be used as the cached key so
that the same result is obtained when jamming equal nouns that use
different integer/binary representations of atoms.

This existing Noun.normalize_noun_atom function (used in the Noun.equal
implementation) is extended to cover cells and renamed to
Noun.normalize_noun. This function is used to normalize nouns when
updating or performing a lookup on the Jam cache.