Fix the TOC logic to properly go up nestings > 1 and dynamic spacing
This patch fixes #555 and fixes #554.

Since spacing on index depends on the number. I.E. if we have a `9.`,
then it should have 3 more spaces, and if it's `10.` it should indent
4 more spaces. We thus keep track of the spacing logic within the
reduce logic.

Further since we have to properly carry around the spacing, this
opened up the issue where the last numbering would not get recomputed
properly. This is a minor issue in reality as livemd would fix the
numbering, however since we must care about the spacing we fix the
logic that is responsible for the incorrect numbering.

Basically the old logic would just take the head of the stack for the
numbering rather than the properly computed one we pass on anyways. So
this is a trivial fix with the new flow