Create a regression test
We create a regression test by repurposing the successful transaction
test. Ideally we'd have examples, and since I didn't want to copy and
paste the test, I just added some logic to it. If we get examples in
the codebase please cut this commit, and reroll this topic for it.

If it's not there, let us include this commit and move it out to a
different example once that work is in