Fix the double spending bug.
This was filed as issue #531. And would allow multiple transactions in
the same block to double spend.

We simply change the flow where we check the nullifier set until after
we can actually write. Blocking for a bit longer.

Note we had to change the flow of logic a bit, as otherwise we would
double wait on getting a `wait_for_ready` message. So some control
flow was added on the failed matches