Move Crypto key generation out of EResource and into ECrypto
As we get more modules into examples, the crypto keys will be used all
over the place, so we move them now.

Instead of making:

1. Example.ECrypto.EId.ex
2. Example.ECrypto.ESymmetric.ex
...

We instead opt to just use ECrypto... EId would be a better name in
part due to less typing, but I believe would muddy the waters name
wise in the future, thus it's ECrypto.