Fix the typing on get keyspace
We do not store the type `qualified_key()`, this referes to a nock
qualification, what we do get back is the key itself with a qualified
value. This key can be any value in practice. It's likely a list of
binaries, for now I've loosend the type to any, but we should maybe
revisit this.

This topic is based off of:

e2adbf538f780ed050f27f0b2de2791d630ec13e

However the bug really appeared in:

57336a5ce8e3e554a6f64ed8966aa1d12e769a01

Due to file changes we can't realistically base the bug fix off of it.