Add the hash of the nullifier to storate rather than the nullifier
This is inline with the commitment hash changes. Much like the
commitment hash changes we should paramartize it and figure out a way
around it, as they are naturally already hashed in the shielded
case.... that or we need to change what we call commitments and
nullifiers in the transparent case.