Add the ability to reuse storage from another node.
Before when we passed storage to a node it would empty it out. This is
unfortunate when it comes to examples as it would undo the hard work
to setup the storage in a certain configuration. This patch fixes it
as now we have storage persisting.

We may want to consider changing some of the engineering around this
as we assume the snapshots share the same name which may cause issues
when they are not the same between the two node setups.