Create a Glossery Module
At first we will fill in the glossary values by hand, like what we can
see in this commit.

However, in the future we should fill it in with `defterm` or some
similar macro that automates this process.

The general idea is that when we have common definitions, we can just
link to the terms here for details. These can be moused over in the
web extraction.