Make delta into it's own type
Delta before was just a map, this is all fine and dandy, however if we
wish to add from_noun and to_noun into the codebase without special
casing, we should make delta it's own data type.

If we don't want to special case the negative numbers in the future,
then we should maybe reconsider this plan, but until then, this ought
to be it's own type.