Change `get_keyspace` Semantics
Previously, when a keyspace gotten has at least one absent key, the
entire query returned nothing. We replace this by instead sorting
through the key after reading the keys from the qualified storage.

We return a list of non-absent keys. If no values are present, we return
an empty list.