Create an ETCP Example
This triggers the death of a tcp socket.

Currently this does not assert any behavior, as we don't have cleanup
logic yet. However this can be a nice example once that is in, and we
can do some assertions.

For now this serves as an easy way to test closing off the socket for
future development