Rename Anoma.Node.Logger to Anoma.Node.EventLogger
Before the Logger would shadow the Elixir.Logger, causing some
annoyances. Now the alias of Anoma.Node.EventLogger would be it's own
unique module, which is a good perk to have

Further we reorder the aliases in the .iex.exs file as it's sorted by
alphabetical order, which should be preserved.

Some other line changes occured due to the length getting longer than
the formatter would want, but no functionality should have been
changed in this commit