event broker: introduce deffilter macro
The deffilter macro permits inline succinct filter definitions. It takes
a module name (which becomes a child of the current module, e.g.
deffilter Trivial inside EventBroker.Filters defines
EventBroker.Filters.Trivial), and a do-block containing the body of the
filter function.

Optional parameters to deffilter are event, a pattern which is used in
the filter function head (defaulting to _), params, working similarly
for the other argument of the filter function head, and fields, a list
of field macros which goes inside the typedstruct defined for the filter
module, defaulting to a fieldless struct.