ci: check for noeol files
This has mysterious properties regarding the presence or absence of !.

What's there now empirically works, but don't trust it.

It's not ideal that it doesn't show which files, but there is no
principled way to check for noeol files, surprisingly enough, and this
is the best we can do consistently.