Make Prove Endpoint Scry Properly
Previously, no scry function was provided for the prove
evaluation. Now, we provide a scry function which checks whether a key
is stored in client storage and if not, sends a read-only request to
the node.

Finally, if the resulting noun is a transaction, we store all app data
marked for storage in the appropriate format.