Nock term conversions
Conversions between Elixir-native Nock `Noun.t()` and
the polynomial representative thereof (`nock_poly_term`).
The conversion to native expects the caller to have
typechecked the polynomial term as representing a Nock
term, while the conversion from native is obliged to
produce a polynomial term which passes that typecheck.
The two conversions are obliged to be inverses (assuming
they are called on valid inputs).