Pad binary data for verify_detached jet to required size
A ed25519 signature is 64 bytes and a verifying key is 32 bytes. These
are passed to the verify_detached jet as atoms. These atoms must be
padded to the expected size otherwise the verification can fail.

For example, as part of the Kudos application test the following
signature was generated:

1e66a254ff3897aa6657e418258b9a8ebc59f165deb17858f18e56e48f1f4bc63519c4d3d0e776835bbcc1a955e6310c11df59de27d907612667063b09a6d0

The last byte of this signature is 0 so when its atom representation is
converted back to bytes by Noun.atom_integer_to_binary, the resulting
binary is only 63 bytes.

The final byte is missing and so the signature verification check fails.