Extended polynomial terms and s-expressions
Define an extended polynomial term type which represents
open polynomial-form Nock terms (which in turn can compile
to internal Elixir Nock nouns by substitution of their
metavariables) with additional constructors to allow us
to define an internal extended language of which Nock
terms remain valid terms, but we may also use other
constructors which can represent operations made available
by Hoon as well as other potential conveniences.

We also define an S-expression Elixir type which can be
parsed to extended open polynomial Nock terms.  This
S-expression representation is more lax than the
polynomial term representation (it accepts some forms,
such as singleton lists, if it can interpret them as
polynomial terms, even though singleton cells are not
themselves valid Nock terms), so it is not an external
interface, but rather an internal convenience -- a compact
way of writing terms that are more verbose if you have to
spell out `:atom` and `:cell` (as well as any other
constructors later added to the extended language).

We start with rather trivial aliases for 0 (:slot), 1
(:constant), and 2 (:evaluate).